Add Directionable module - include in RoadTripsCreateFacade & MunchiesIndexFacade
